genericOpenFigiAPI.py
This code is for mapping from third-party market data identifiers to Bloomberg Tickers. Accepts the type of job to run based on the API specification (https://www.openfigi.com/api). Initial cut requires mapping jobs to be manually added to the script, although I will make this an input soon.